+++ pkg/planner/core/indexmerge_path.go
@@ -43,97 +43,6 @@ import (
 	"go.uber.org/zap"
 )
 
-// generateIndexMergePath generates IndexMerge AccessPaths on this DataSource.
-func generateIndexMergePath(ds *logicalop.DataSource) error {
-	if ds.SCtx().GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
-		debugtrace.EnterContextCommon(ds.SCtx())
-		defer debugtrace.LeaveContextCommon(ds.SCtx())
-	}
-	var warningMsg string
-	stmtCtx := ds.SCtx().GetSessionVars().StmtCtx
-	defer func() {
-		if len(ds.IndexMergeHints) > 0 && warningMsg != "" {
-			ds.IndexMergeHints = nil
-			stmtCtx.AppendWarning(errors.NewNoStackError(warningMsg))
-			logutil.BgLogger().Debug(warningMsg)
-		}
-	}()
-
-	// Consider the IndexMergePath. Now, we just generate `IndexMergePath` in DNF case.
-	// Use AllConds instread of PushedDownConds,
-	// because we want to use IndexMerge even if some expr cannot be pushed to TiKV.
-	// We will create new Selection for exprs that cannot be pushed in convertToIndexMergeScan.
-	indexMergeConds := make([]expression.Expression, 0, len(ds.AllConds))
-	indexMergeConds = append(indexMergeConds, ds.AllConds...)
-	sessionAndStmtPermission := (ds.SCtx().GetSessionVars().GetEnableIndexMerge() || len(ds.IndexMergeHints) > 0) && !stmtCtx.NoIndexMergeHint
-	if !sessionAndStmtPermission {
-		warningMsg = "IndexMerge is inapplicable or disabled. Got no_index_merge hint or tidb_enable_index_merge is off."
-		return nil
-	}
-
-	if ds.TableInfo.TempTableType == model.TempTableLocal {
-		warningMsg = "IndexMerge is inapplicable or disabled. Cannot use IndexMerge on temporary table."
-		return nil
-	}
-
-	regularPathCount := len(ds.PossibleAccessPaths)
-
-	// Now we have 3 entry functions to generate IndexMerge paths:
-	// 1. Generate AND type IndexMerge for non-MV indexes and all OR type IndexMerge.
-	var err error
-	if warningMsg, err = generateOtherIndexMerge(ds, regularPathCount, indexMergeConds); err != nil {
-		return err
-	}
-	// 2. Generate AND type IndexMerge for MV indexes. Tt can only use one index in an IndexMerge path.
-	if err := generateANDIndexMerge4MVIndex(ds, regularPathCount, indexMergeConds); err != nil {
-		return err
-	}
-	oldIndexMergeCount := len(ds.PossibleAccessPaths)
-	// 3. Generate AND type IndexMerge for MV indexes. It can use multiple MV and non-MV indexes in an IndexMerge path.
-	if err := generateANDIndexMerge4ComposedIndex(ds, regularPathCount, indexMergeConds); err != nil {
-		return err
-	}
-
-	// Because index merge access paths are built from ds.AllConds, sometimes they can help us consider more filters than
-	// the ds.stats, which is calculated from ds.PushedDownConds before this point.
-	// So we use a simple and naive method to update ds.stats here using the largest row count from index merge paths.
-	// This can help to avoid some cases where the row count of operator above IndexMerge is larger than IndexMerge.
-	// TODO: Probably we should directly consider ds.AllConds when calculating ds.stats in the future.
-	if len(ds.PossibleAccessPaths) > regularPathCount && len(ds.AllConds) > len(ds.PushedDownConds) {
-		var maxRowCount float64
-		for i := regularPathCount; i < len(ds.PossibleAccessPaths); i++ {
-			maxRowCount = max(maxRowCount, ds.PossibleAccessPaths[i].CountAfterAccess)
-		}
-		if ds.StatsInfo().RowCount > maxRowCount {
-			ds.SetStats(ds.TableStats.ScaleByExpectCnt(ds.SCtx().GetSessionVars(), maxRowCount))
-		}
-	}
-
-	// If without hints, it means that `enableIndexMerge` is true
-	if len(ds.IndexMergeHints) != 0 {
-		// If len(IndexMergeHints) > 0, then add warnings if index-merge hints cannot work.
-		if regularPathCount == len(ds.PossibleAccessPaths) {
-			if warningMsg == "" {
-				warningMsg = "IndexMerge is inapplicable"
-			}
-			return nil
-		}
-
-		// If len(IndexMergeHints) > 0 and some index-merge paths were added, then prune all other non-index-merge paths.
-		// if len(ds.PossibleAccessPaths) > oldIndexMergeCount, it means composed index merge path is generated, prune others.
-		if len(ds.PossibleAccessPaths) > oldIndexMergeCount {
-			ds.PossibleAccessPaths = ds.PossibleAccessPaths[oldIndexMergeCount:]
-		} else {
-			ds.PossibleAccessPaths = ds.PossibleAccessPaths[regularPathCount:]
-		}
-	}
-
-	// If there is a multi-valued index hint, remove all paths which don't use the specified index.
-	cleanAccessPathForMVIndexHint(ds)
-
-	return nil
-}
-
 func generateNormalIndexPartialPath(
 	ds *logicalop.DataSource,
 	item expression.Expression,
@@ -171,34 +80,6 @@ func isInIndexMergeHints(ds *logicalop.DataSource, name string) bool {
 	for _, hint := range ds.IndexMergeHints {
 		if hint.IndexHint == nil || len(hint.IndexHint.IndexNames) == 0 {
 			return true
-		}
-		for _, hintName := range hint.IndexHint.IndexNames {
-			if strings.EqualFold(name, hintName.String()) {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// indexMergeHintsHasSpecifiedIdx returns true if there's IndexMerge hint, and it has specified index names.
-func indexMergeHintsHasSpecifiedIdx(ds *logicalop.DataSource) bool {
-	for _, hint := range ds.IndexMergeHints {
-		if hint.IndexHint == nil || len(hint.IndexHint.IndexNames) == 0 {
-			continue
-		}
-		if len(hint.IndexHint.IndexNames) > 0 {
-			return true
-		}
-	}
-	return false
-}
-
-// indexMergeHintsHasSpecifiedIdx return true if the input index name is specified in the IndexMerge hint.
-func isSpecifiedInIndexMergeHints(ds *logicalop.DataSource, name string) bool {
-	for _, hint := range ds.IndexMergeHints {
-		if hint.IndexHint == nil || len(hint.IndexHint.IndexNames) == 0 {
-			continue
 		}
 		for _, hintName := range hint.IndexHint.IndexNames {
 			if strings.EqualFold(name, hintName.String()) {
@@ -265,143 +146,6 @@ func generateNormalIndexPartialPath4And(ds *logicalop.DataSource, normalPathCnt 
 		return res.PartialIndexPaths
 	}
 	return nil
-}
-
-// generateANDIndexMerge4NormalIndex generates IndexMerge paths for `AND` (a.k.a. intersection type IndexMerge)
-func generateANDIndexMerge4NormalIndex(ds *logicalop.DataSource, normalPathCnt int, usedAccessMap map[string]expression.Expression) *util.AccessPath {
-	// For now, we only consider intersection type IndexMerge when the index names are specified in the hints.
-	if !indexMergeHintsHasSpecifiedIdx(ds) {
-		return nil
-	}
-	composedWithMvIndex := len(usedAccessMap) != 0
-
-	// 1. Collect partial paths from normal paths.
-	partialPaths := make([]*util.AccessPath, 0, normalPathCnt)
-	for i := range normalPathCnt {
-		originalPath := ds.PossibleAccessPaths[i]
-		// No need to consider table path as a partial path.
-		if ds.PossibleAccessPaths[i].IsTablePath() {
-			continue
-		}
-		// since this code path is only for normal index, skip mv index here.
-		if ds.PossibleAccessPaths[i].Index.MVIndex {
-			continue
-		}
-		if !isSpecifiedInIndexMergeHints(ds, originalPath.Index.Name.L) {
-			continue
-		}
-		// If the path contains a full range, ignore it.
-		if ranger.HasFullRange(originalPath.Ranges, false) {
-			continue
-		}
-		if composedWithMvIndex {
-			// case:
-			// idx1: mv(c, cast(`a` as signed array))
-			// idx2: idx(c)
-			// idx3: idx(c, d)
-			// for condition: (1 member of a) AND (c = 1) AND (d = 2), we should pick idx1 and idx3,
-			// since idx2's access cond has already been covered by idx1.
-			containRelation := true
-			for _, access := range originalPath.AccessConds {
-				if _, ok := usedAccessMap[string(access.HashCode())]; !ok {
-					// some condition is not covered in previous mv index partial path, use it!
-					containRelation = false
-					break
-				}
-			}
-			if containRelation {
-				continue
-			}
-			// for this picked normal index, mark its access conds.
-			for _, access := range originalPath.AccessConds {
-				if _, ok := usedAccessMap[string(access.HashCode())]; !ok {
-					usedAccessMap[string(access.HashCode())] = access
-				}
-			}
-		}
-		newPath := originalPath.Clone()
-		partialPaths = append(partialPaths, newPath)
-	}
-	if len(partialPaths) < 1 {
-		return nil
-	}
-	if len(partialPaths) == 1 && !composedWithMvIndex {
-		// even single normal index path here, it can be composed with other mv index partial paths.
-		return nil
-	}
-
-	// 2. Collect filters that can't be covered by the partial paths and deduplicate them.
-	finalFilters := make([]expression.Expression, 0)
-	partialFilters := make([]expression.Expression, 0, len(partialPaths))
-	hashCodeSet := make(map[string]struct{})
-	pushDownCtx := util.GetPushDownCtx(ds.SCtx())
-	for _, path := range partialPaths {
-		// Classify filters into coveredConds and notCoveredConds.
-		coveredConds := make([]expression.Expression, 0, len(path.AccessConds)+len(path.IndexFilters))
-		notCoveredConds := make([]expression.Expression, 0, len(path.IndexFilters)+len(path.TableFilters))
-		// AccessConds can be covered by partial path.
-		coveredConds = append(coveredConds, path.AccessConds...)
-		path.IndexFilters = slices.DeleteFunc(path.IndexFilters, func(cond expression.Expression) bool {
-			// IndexFilters can be covered by partial path if it can be pushed down to TiKV.
-			if !expression.CanExprsPushDown(pushDownCtx, []expression.Expression{cond}, kv.TiKV) {
-				notCoveredConds = append(notCoveredConds, cond)
-				return true
-			}
-			coveredConds = append(coveredConds, cond)
-			return false
-		})
-		// TableFilters can't be covered by partial path.
-		notCoveredConds = append(notCoveredConds, path.TableFilters...)
-
-		// Record covered filters in hashCodeSet.
-		// Note that we only record filters that not appear in the notCoveredConds. It's possible that a filter appear
-		// in both coveredConds and notCoveredConds (e.g. because of prefix index). So we need this extra check to
-		// avoid wrong deduplication.
-		notCoveredHashCodeSet := make(map[string]struct{})
-		for _, cond := range notCoveredConds {
-			hashCode := string(cond.HashCode())
-			notCoveredHashCodeSet[hashCode] = struct{}{}
-		}
-		for _, cond := range coveredConds {
-			hashCode := string(cond.HashCode())
-			if _, ok := notCoveredHashCodeSet[hashCode]; !ok {
-				hashCodeSet[hashCode] = struct{}{}
-			}
-		}
-
-		finalFilters = append(finalFilters, notCoveredConds...)
-		partialFilters = append(partialFilters, coveredConds...)
-	}
-
-	// Remove covered filters from finalFilters and deduplicate finalFilters.
-	dedupedFinalFilters := make([]expression.Expression, 0, len(finalFilters))
-	for _, cond := range finalFilters {
-		hashCode := string(cond.HashCode())
-		if _, ok := hashCodeSet[hashCode]; !ok {
-			dedupedFinalFilters = append(dedupedFinalFilters, cond)
-			hashCodeSet[hashCode] = struct{}{}
-		}
-	}
-
-	// Keep these partial filters as a part of table filters for safety if there is any parameter.
-	if expression.MaybeOverOptimized4PlanCache(ds.SCtx().GetExprCtx(), partialFilters...) {
-		dedupedFinalFilters = append(dedupedFinalFilters, partialFilters...)
-	}
-
-	// 3. Estimate the row count after partial paths.
-	sel, _, err := cardinality.Selectivity(ds.SCtx(), ds.TableStats.HistColl, partialFilters, nil)
-	if err != nil {
-		logutil.BgLogger().Debug("something wrong happened, use the default selectivity", zap.Error(err))
-		sel = cost.SelectionFactor
-	}
-
-	indexMergePath := &util.AccessPath{
-		PartialIndexPaths:        partialPaths,
-		IndexMergeIsIntersection: true,
-		TableFilters:             dedupedFinalFilters,
-		CountAfterAccess:         sel * ds.TableStats.RowCount,
-	}
-	return indexMergePath
 }
 
 // generateMVIndexMergePartialPaths4And try to find mv index merge partial path from a collection of cnf conditions.
@@ -507,84 +251,6 @@ func generateMVIndexMergePartialPaths4And(ds *logicalop.DataSource, normalPathCn
 		mvAndPartialPath = append(mvAndPartialPath, one.paths...)
 	}
 	return mvAndPartialPath, usedAccessCondsMap, nil
-}
-
-// generateOtherIndexMerge is the entry point for generateORIndexMerge() and generateANDIndexMerge4NormalIndex(), plus
-// some extra logic to keep some specific behaviors the same as before.
-func generateOtherIndexMerge(ds *logicalop.DataSource, regularPathCount int, indexMergeConds []expression.Expression) (string, error) {
-	isPossibleIdxMerge := len(indexMergeConds) > 0 && // have corresponding access conditions, and
-		len(ds.PossibleAccessPaths) > 1 // have multiple index paths
-	if !isPossibleIdxMerge {
-		return "IndexMerge is inapplicable or disabled. No available filter or available index.", nil
-	}
-
-	// We current do not consider `IndexMergePath`:
-	// 1. If there is an index path.
-	// 2. TODO: If there exists exprs that cannot be pushed down. This is to avoid wrongly estRow of Selection added by rule_predicate_push_down.
-	stmtCtx := ds.SCtx().GetSessionVars().StmtCtx
-	needConsiderIndexMerge := true
-	// if current index merge hint is nil, once there is a no-access-cond in one of possible access path.
-	if len(ds.IndexMergeHints) == 0 {
-		skipRangeScanCheck := fixcontrol.GetBoolWithDefault(
-			ds.SCtx().GetSessionVars().GetOptimizerFixControlMap(),
-			fixcontrol.Fix52869,
-			false,
-		)
-		ds.SCtx().GetSessionVars().RecordRelevantOptFix(fixcontrol.Fix52869)
-		if !skipRangeScanCheck {
-			for i := 1; i < len(ds.PossibleAccessPaths); i++ {
-				if len(ds.PossibleAccessPaths[i].AccessConds) != 0 {
-					needConsiderIndexMerge = false
-					break
-				}
-			}
-		}
-		if needConsiderIndexMerge {
-			// PushDownExprs() will append extra warnings, which is annoying. So we reset warnings here.
-			warnings := stmtCtx.GetWarnings()
-			extraWarnings := stmtCtx.GetExtraWarnings()
-			_, remaining := expression.PushDownExprs(util.GetPushDownCtx(ds.SCtx()), indexMergeConds, kv.UnSpecified)
-			stmtCtx.SetWarnings(warnings)
-			stmtCtx.SetExtraWarnings(extraWarnings)
-			if len(remaining) > 0 {
-				needConsiderIndexMerge = false
-			}
-		}
-	}
-
-	// 1. Generate possible IndexMerge paths for `OR`.
-	err := generateORIndexMerge(ds, indexMergeConds)
-	if err != nil {
-		return "", err
-	}
-	// 2. Generate possible IndexMerge paths for `AND`.
-	indexMergeAndPath := generateANDIndexMerge4NormalIndex(ds, regularPathCount, nil)
-	if indexMergeAndPath != nil {
-		ds.PossibleAccessPaths = append(ds.PossibleAccessPaths, indexMergeAndPath)
-	}
-
-	if needConsiderIndexMerge {
-		return "", nil
-	}
-
-	var containMVPath bool
-	for i := regularPathCount; i < len(ds.PossibleAccessPaths); i++ {
-		path := ds.PossibleAccessPaths[i]
-		for _, p := range util.SliceRecursiveFlattenIter[*util.AccessPath](path.PartialAlternativeIndexPaths) {
-			if isMVIndexPath(p) {
-				containMVPath = true
-				break
-			}
-		}
-
-		if !containMVPath {
-			ds.PossibleAccessPaths = slices.Delete(ds.PossibleAccessPaths, i, i+1)
-		}
-	}
-	if len(ds.PossibleAccessPaths) == regularPathCount {
-		return "IndexMerge is inapplicable or disabled. ", nil
-	}
-	return "", nil
 }
 
 // generateANDIndexMerge4ComposedIndex tries to generate AND type index merge AccessPath for ( json_member_of /
