+++ pkg/planner/core/stats/util.go
@@ -1,0 +1,136 @@
+// Copyright 2025 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package stats
+
+import (
+	"slices"
+
+	"github.com/pingcap/tidb/pkg/expression"
+	"github.com/pingcap/tidb/pkg/meta/model"
+	"github.com/pingcap/tidb/pkg/planner/cardinality"
+	"github.com/pingcap/tidb/pkg/planner/core/cost"
+	"github.com/pingcap/tidb/pkg/planner/core/operator/logicalop"
+	"github.com/pingcap/tidb/pkg/planner/property"
+	"github.com/pingcap/tidb/pkg/planner/util"
+	"github.com/pingcap/tidb/pkg/planner/util/debugtrace"
+	statsutil "github.com/pingcap/tidb/pkg/planner/util/stats"
+	"github.com/pingcap/tidb/pkg/sessionctx/stmtctx"
+	"github.com/pingcap/tidb/pkg/statistics"
+	"github.com/pingcap/tidb/pkg/util/logutil"
+	"go.uber.org/zap"
+)
+
+func InitStats(ds *logicalop.DataSource) {
+	if ds.StatisticTable == nil {
+		ds.StatisticTable = statsutil.GetStatsTable(ds.SCtx(), ds.TableInfo, ds.PhysicalTableID)
+	}
+	tableStats := &property.StatsInfo{
+		RowCount:     float64(ds.StatisticTable.RealtimeCount),
+		ColNDVs:      make(map[int64]float64, ds.Schema().Len()),
+		HistColl:     ds.StatisticTable.GenerateHistCollFromColumnInfo(ds.TableInfo, ds.TblCols),
+		StatsVersion: ds.StatisticTable.Version,
+	}
+	if ds.StatisticTable.Pseudo {
+		tableStats.StatsVersion = statistics.PseudoVersion
+	}
+
+	statsRecord := ds.SCtx().GetSessionVars().StmtCtx.GetUsedStatsInfo(true)
+	name, tblInfo := statsutil.GetTblInfoForUsedStatsByPhysicalID(ds.SCtx(), ds.PhysicalTableID)
+	statsRecord.RecordUsedInfo(ds.PhysicalTableID, &stmtctx.UsedStatsInfoForTable{
+		Name:            name,
+		TblInfo:         tblInfo,
+		Version:         tableStats.StatsVersion,
+		RealtimeCount:   tableStats.HistColl.RealtimeCount,
+		ModifyCount:     tableStats.HistColl.ModifyCount,
+		ColAndIdxStatus: ds.StatisticTable.ColAndIdxExistenceMap,
+	})
+
+	for _, col := range ds.Schema().Columns {
+		tableStats.ColNDVs[col.UniqueID] = cardinality.EstimateColumnNDV(ds.StatisticTable, col.ID)
+	}
+	ds.TableStats = tableStats
+	ds.TableStats.GroupNDVs = getGroupNDVs(ds)
+	ds.TblColHists = ds.StatisticTable.ID2UniqueID(ds.TblCols)
+	for _, col := range ds.TableInfo.Columns {
+		if col.State != model.StatePublic {
+			continue
+		}
+	}
+}
+
+func DeriveStatsByFilter(ds *logicalop.DataSource, conds expression.CNFExprs, filledPaths []*util.AccessPath) *property.StatsInfo {
+	if ds.SCtx().GetSessionVars().StmtCtx.EnableOptimizerDebugTrace {
+		debugtrace.EnterContextCommon(ds.SCtx())
+		defer debugtrace.LeaveContextCommon(ds.SCtx())
+	}
+	selectivity, _, err := cardinality.Selectivity(ds.SCtx(), ds.TableStats.HistColl, conds, filledPaths)
+	if err != nil {
+		logutil.BgLogger().Debug("something wrong happened, use the default selectivity", zap.Error(err))
+		selectivity = cost.SelectionFactor
+	}
+	// TODO: remove NewHistCollBySelectivity later on.
+	// if ds.SCtx().GetSessionVars().OptimizerSelectivityLevel >= 1 {
+	// Only '0' is suggested, see https://docs.pingcap.com/zh/tidb/stable/system-variables#tidb_optimizer_selectivity_level.
+	// stats.HistColl = stats.HistColl.NewHistCollBySelectivity(ds.SCtx(), nodes)
+	// }
+	return ds.TableStats.Scale(ds.SCtx().GetSessionVars(), selectivity)
+}
+
+func getGroupNDVs(ds *logicalop.DataSource) []property.GroupNDV {
+	colGroups := ds.AskedColumnGroup
+	if len(ds.AskedColumnGroup) == 0 {
+		return nil
+	}
+	tbl := ds.TableStats.HistColl
+	ndvs := make([]property.GroupNDV, 0, len(colGroups))
+	tbl.ForEachIndexImmutable(func(idxID int64, idx *statistics.Index) bool {
+		colsLen := len(tbl.Idx2ColUniqueIDs[idxID])
+		// tbl.Idx2ColUniqueIDs may only contain the prefix of index columns.
+		// But it may exceeds the total index since the index would contain the handle column if it's not a unique index.
+		// We append the handle at fillIndexPath.
+		if colsLen < len(idx.Info.Columns) {
+			return false
+		} else if colsLen > len(idx.Info.Columns) {
+			colsLen--
+		}
+		idxCols := make([]int64, colsLen)
+		copy(idxCols, tbl.Idx2ColUniqueIDs[idxID])
+		slices.Sort(idxCols)
+		for _, g := range colGroups {
+			// We only want those exact matches.
+			if len(g) != colsLen {
+				return false
+			}
+			match := true
+			for i, col := range g {
+				// Both slices are sorted according to UniqueID.
+				if col.UniqueID != idxCols[i] {
+					match = false
+					break
+				}
+			}
+			if match && idx.IsEssentialStatsLoaded() {
+				ndv := property.GroupNDV{
+					Cols: idxCols,
+					NDV:  float64(idx.NDV),
+				}
+				ndvs = append(ndvs, ndv)
+				return true
+			}
+		}
+		return false
+	})
+	return ndvs
+}
